+++ app/controllers/folio/ui_controller.rb
@@ -1,7 +1,8 @@
 # frozen_string_literal: true
 
 class Folio::UiController < Folio::BaseController
-  before_action :authenticate_account!
+  before_action :authenticate_user!
+  before_action :authorize_user!
 
   def ui
   end
@@ -10,4 +11,12 @@ class Folio::UiController < Folio::BaseController
     @mobile_typo = true
     render :ui
   end
+
+  def atoms
+  end
+
+  private
+    def authorize_user!
+      raise CanCan::AccessDenied unless can_now?(:display_ui)
+    end
 end