+++ config/initializers/omniauth.rb
@@ -2,10 +2,8 @@
 
 require Folio::Engine.root.join("app/models/folio/omniauth") # to load Folio::Omniauth namespace
 
-if Rails.application.config.folio_users
-  Folio::Omniauth.setup_providers(Rails.application.config.folio_users_omniauth_providers)
+Folio::Omniauth.setup_providers(Rails.application.config.folio_users_omniauth_providers)
 
-  Rails.application.config.action_dispatch.cookies_same_site_protection = lambda { |request|
-    request.path == "/users/auth/apple" ? :none : :lax
-  }
-end
+Rails.application.config.action_dispatch.cookies_same_site_protection = lambda { |request|
+  request.path == "/users/auth/apple" ? :none : :lax
+}