+++ test/controllers/folio/api/s3_controller_test.rb
@@ -8,8 +8,8 @@ class Folio::Api::S3ControllerTest < ActionDispatch::IntegrationTest
 
   def setup
     create_and_host_site
-    @admin = create(:folio_account)
-    sign_in @admin
+    @superadmin = create(:folio_user, :superadmin)
+    sign_in @superadmin
   end
 
   [Folio::File::Document, Folio::File::Image, Folio::PrivateAttachment].each do |klass|
@@ -70,8 +70,8 @@ class Folio::Api::S3ControllerTest < ActionDispatch::IntegrationTest
       assert_equal("test-#{klass.model_name.singular}-#{file.id}.gif", file.reload.file_name)
     end
 
-    test "#{klass} - unauthorized for no account" do
-      sign_out @admin
+    test "#{klass} - unauthorized for no admin" do
+      sign_out @superadmin
 
       post before_folio_api_s3_path, params: { file_name: "Intricate fílě name.jpg" }
       assert_response :unauthorized