+++ test/controllers/folio/console/dashboard_controller_test.rb
@@ -10,10 +10,11 @@ module Folio
       get console_root_url
       assert_redirected_to console_pages_path
 
-      @admin.forget_me!
-      sign_out @admin
+      superadmin.forget_me!
+      sign_out superadmin
+
       get console_root_url
-      assert_redirected_to new_account_session_path
+      assert_redirected_to new_user_session_path
     end
   end
 end