+++ test/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2023_11_06_080839) do
+ActiveRecord::Schema[7.0].define(version: 2023_12_19_123901) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_trgm"
   enable_extension "plpgsql"
@@ -91,45 +91,6 @@ ActiveRecord::Schema[7.0].define(version: 2023_11_06_080839) do
     t.index ["position"], name: "index_dummy_blog_topics_on_position"
     t.index ["published"], name: "index_dummy_blog_topics_on_published"
     t.index ["slug"], name: "index_dummy_blog_topics_on_slug"
-  end
-
-  create_table "folio_accounts", force: :cascade do |t|
-    t.string "email", default: "", null: false
-    t.string "encrypted_password", default: "", null: false
-    t.string "reset_password_token"
-    t.datetime "reset_password_sent_at", precision: nil
-    t.datetime "remember_created_at", precision: nil
-    t.integer "sign_in_count", default: 0, null: false
-    t.datetime "current_sign_in_at", precision: nil
-    t.datetime "last_sign_in_at", precision: nil
-    t.string "current_sign_in_ip"
-    t.string "last_sign_in_ip"
-    t.datetime "created_at", precision: nil, null: false
-    t.datetime "updated_at", precision: nil, null: false
-    t.string "first_name"
-    t.string "last_name"
-    t.boolean "is_active", default: true
-    t.string "invitation_token"
-    t.datetime "invitation_created_at", precision: nil
-    t.datetime "invitation_sent_at", precision: nil
-    t.datetime "invitation_accepted_at", precision: nil
-    t.integer "invitation_limit"
-    t.string "invited_by_type"
-    t.bigint "invited_by_id"
-    t.integer "invitations_count", default: 0
-    t.string "crossdomain_devise_token"
-    t.datetime "crossdomain_devise_set_at"
-    t.string "sign_out_salt_part"
-    t.jsonb "roles", default: []
-    t.string "console_path"
-    t.datetime "console_path_updated_at"
-    t.index ["crossdomain_devise_token"], name: "index_folio_accounts_on_crossdomain_devise_token"
-    t.index ["email"], name: "index_folio_accounts_on_email", unique: true
-    t.index ["invitation_token"], name: "index_folio_accounts_on_invitation_token", unique: true
-    t.index ["invitations_count"], name: "index_folio_accounts_on_invitations_count"
-    t.index ["invited_by_id"], name: "index_folio_accounts_on_invited_by_id"
-    t.index ["invited_by_type", "invited_by_id"], name: "index_folio_accounts_on_invited_by"
-    t.index ["reset_password_token"], name: "index_folio_accounts_on_reset_password_token", unique: true
   end
 
   create_table "folio_addresses", force: :cascade do |t|
@@ -417,7 +378,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_11_06_080839) do
   create_table "folio_site_user_links", force: :cascade do |t|
     t.bigint "user_id", null: false
     t.bigint "site_id", null: false
-    t.json "roles", default: []
+    t.jsonb "roles", default: []
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["site_id"], name: "index_folio_site_user_links_on_site_id"
@@ -450,7 +411,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_11_06_080839) do
     t.string "slug"
     t.integer "position"
     t.string "copyright_info_source"
-    t.json "available_user_roles", default: []
+    t.jsonb "available_user_roles", default: ["administrator", "manager"]
     t.index ["domain"], name: "index_folio_sites_on_domain"
     t.index ["position"], name: "index_folio_sites_on_position"
     t.index ["slug"], name: "index_folio_sites_on_slug"
@@ -496,6 +457,9 @@ ActiveRecord::Schema[7.0].define(version: 2023_11_06_080839) do
     t.datetime "crossdomain_devise_set_at"
     t.string "sign_out_salt_part"
     t.bigint "source_site_id"
+    t.boolean "superadmin", default: false, null: false
+    t.string "console_path"
+    t.datetime "console_path_updated_at"
     t.index ["confirmation_token"], name: "index_folio_users_on_confirmation_token", unique: true
     t.index ["crossdomain_devise_token"], name: "index_folio_users_on_crossdomain_devise_token"
     t.index ["email"], name: "index_folio_users_on_email"